code refactor, created InventoryReservationCli
