+++ app/code/Magento/InventoryReservationCli/composer.json
@@ -1,0 +1,33 @@
+{
+    "name": "magento/inventoryreservationcli",
+    "description": "",
+    "require": {
+        "php": "~5.5.0|~5.6.0|~7.0.0",
+        
+        "magento/magento-composer-installer": "*"
+    },
+    "suggest": {
+
+    },
+    "type": "magento2-module",
+    "version": "0.1.0",
+    "license": [
+
+    ],
+    "autoload": {
+        "files": [
+            "registration.php"
+        ],
+        "psr-4": {
+            "Magento\\InventoryReservationCli\\": ""
+        }
+    },
+    "extra": {
+        "map": [
+            [
+                "*",
+                "Magento/InventoryReservationCli"
+            ]
+        ]
+    }
+}
