+++ app/code/Magento/InventoryReservationCli/etc/di.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0"?>
+<!--
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
+    <type name="Magento\Framework\Console\CommandList">
+        <arguments>
+            <argument name="commands" xsi:type="array">
+                <item name="inventory_complete_order_inconsistency" xsi:type="object">Magento\InventoryReservationCli\Command\ShowInconsistencyOrderComplete</item>
+            </argument>
+        </arguments>
+    </type>
+</config>
