+++ app/code/Magento/InventoryReservationCli/etc/module.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
+    <module name="Magento_InventoryReservationCli" setup_version="0.1.0">
+        <sequence>
+            <module name="Magento_InventoryReservations"/>
+            <module name="Magento_Sales"/>
+
+        </sequence>
+    </module>
+</config>
