+++ app/code/Magento/InventoryReservationCli/registration.php
@@ -1,0 +1,7 @@
+<?php
+
+\Magento\Framework\Component\ComponentRegistrar::register(
+    \Magento\Framework\Component\ComponentRegistrar::MODULE,
+    'Magento_InventoryReservationCli',
+    __DIR__
+);
